+++ src/Particle/Par_FindHomePatch_UniformGrid.cpp
@@ -70,9 +70,9 @@ void Par_FindHomePatch_UniformGrid( const int lv, const bool OldParOnly,
    real TParPos[3];
 
    long *HomeLBIdx          = new long [NTarPar];
-   int  *HomeLBIdx_IdxTable = new int  [NTarPar];
+   long *HomeLBIdx_IdxTable = new long [NTarPar];
    int  *HomePID            = new int  [NTarPar];
-   int  *MatchIdx           = new int  [NTarPar];
+   long *MatchIdx           = new long [NTarPar];
 
 // get the load-balance index of the particle's home patch
    for (long t=0; t<NTarPar; t++)
@@ -94,13 +94,13 @@ void Par_FindHomePatch_UniformGrid( const int lv, const bool OldParOnly,
    Mis_Heapsort( NReal, RealPatchLBIdx, RealPatchLBIdx_IdxTable );
 
 // LBIdx --> home (real) patch indices
-   Mis_Matching_int( NReal, RealPatchLBIdx, NTarPar, HomeLBIdx, MatchIdx );
+   Mis_Matching_int( (long)NReal, RealPatchLBIdx, NTarPar, HomeLBIdx, MatchIdx );
 
 // check: every particle must have a home patch
 #  ifdef DEBUG_PARTICLE
    for (long t=0; t<NTarPar; t++)
    {
-      if ( MatchIdx[t] == -1 )
+      if ( MatchIdx[t] == -1L )
       {
          const long ParID = NewParID0 + HomeLBIdx_IdxTable[t];
 
@@ -201,7 +201,7 @@ void SendParticle2HomeRank( const int lv, const bool OldParOnly,
    real *Mass   = NULL;
    real *Pos[3] = { NULL, NULL, NULL };
 
-   int  Send_Count[MPI_NRank], Recv_Count[MPI_NRank];
+   long Send_Count[MPI_NRank], Recv_Count[MPI_NRank];
    long Send_Disp[MPI_NRank], Recv_Disp[MPI_NRank];
    long Send_Count_Sum, Recv_Count_Sum;
 
@@ -221,7 +221,7 @@ void SendParticle2HomeRank( const int lv, const bool OldParOnly,
       Pos[2] = NewParAtt[PAR_POSZ];
    }
 
-   for (int r=0; r<MPI_NRank; r++)  Send_Count[r] = 0;
+   for (int r=0; r<MPI_NRank; r++)  Send_Count[r] = 0L;
 
 
 // 1. get the target MPI rank of each particle
@@ -255,19 +255,19 @@ void SendParticle2HomeRank( const int lv, const bool OldParOnly,
 
 
 // 2. construct the MPI send and recv data list
-   MPI_Alltoall( Send_Count, 1, MPI_INT, Recv_Count, 1, MPI_INT, MPI_COMM_WORLD );
+   MPI_Alltoall( Send_Count, 1, MPI_LONG, Recv_Count, 1, MPI_LONG, MPI_COMM_WORLD );
 
    Send_Disp[0] = 0L;
    Recv_Disp[0] = 0L;
 
    for (int r=1; r<MPI_NRank; r++)
    {
-      Send_Disp[r] = Send_Disp[r-1] + (long)Send_Count[r-1];
-      Recv_Disp[r] = Recv_Disp[r-1] + (long)Recv_Count[r-1];
+      Send_Disp[r] = Send_Disp[r-1] + Send_Count[r-1];
+      Recv_Disp[r] = Recv_Disp[r-1] + Recv_Count[r-1];
    }
 
-   Send_Count_Sum = Send_Disp[ MPI_NRank-1 ] + (long)Send_Count[ MPI_NRank-1 ];
-   Recv_Count_Sum = Recv_Disp[ MPI_NRank-1 ] + (long)Recv_Count[ MPI_NRank-1 ];
+   Send_Count_Sum = Send_Disp[ MPI_NRank-1 ] + Send_Count[ MPI_NRank-1 ];
+   Recv_Count_Sum = Recv_Disp[ MPI_NRank-1 ] + Recv_Count[ MPI_NRank-1 ];
 
 #  ifdef DEBUG_PARTICLE
    if ( OldParOnly  &&  Send_Count_Sum != amr->Par->NPar_Active )
@@ -327,8 +327,8 @@ void SendParticle2HomeRank( const int lv, const bool OldParOnly,
    {
       free( amr->Par->InactiveParList );
 
-      amr->Par->NPar_AcPlusInac     = (long)Recv_Count_Sum;
-      amr->Par->NPar_Active         = (long)Recv_Count_Sum;
+      amr->Par->NPar_AcPlusInac     = Recv_Count_Sum;
+      amr->Par->NPar_Active         = Recv_Count_Sum;
       amr->Par->NPar_Inactive       = 0;                       // since we don't redistribute inactive particles
       amr->Par->ParListSize         = UpdatedParListSize;
       amr->Par->InactiveParListSize = (long)MAX( 1, UpdatedParListSize/100 );
@@ -337,8 +337,8 @@ void SendParticle2HomeRank( const int lv, const bool OldParOnly,
 
    else
    {
-      amr->Par->NPar_AcPlusInac    += (long)Recv_Count_Sum;
-      amr->Par->NPar_Active        += (long)Recv_Count_Sum;
+      amr->Par->NPar_AcPlusInac    += Recv_Count_Sum;
+      amr->Par->NPar_Active        += Recv_Count_Sum;
       amr->Par->ParListSize         = UpdatedParListSize;
 
 //    update the total number of active particles in all MPI ranks
